studing/requirements.txt
Django==3.1.7
django-rest-framework
black
isort==4.2.5
pytest==5.4.3
pytest-isort==1.1.0
pytest-black==0.3.10
pytest-cov==2.10.0
pytest-django==3.8.0
pytest-mock